+++ packages/ui/src/components/Alert/SAlert.spec.cy.ts
